extract projection into subpackage
