+++ internal/handlers/common/projection/projection_iterator.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package common
+package projection
 
 import (
 	"errors"
