+++ internal/handlers/tigris/msg_find.go
@@ -20,6 +20,7 @@ import (
 	"time"
 
 	"github.com/FerretDB/FerretDB/internal/handlers/common"
+	"github.com/FerretDB/FerretDB/internal/handlers/common/projection"
 	"github.com/FerretDB/FerretDB/internal/handlers/commonerrors"
 	"github.com/FerretDB/FerretDB/internal/handlers/tigris/tigrisdb"
 	"github.com/FerretDB/FerretDB/internal/types"
@@ -100,7 +101,7 @@ func (h *Handler) MsgFind(ctx context.Context, msg *wire.OpMsg) (*wire.OpMsg, er
 
 		iter = common.LimitIterator(iter, closer, params.Limit)
 
-		iter, err = common.ProjectionIterator(iter, closer, params.Projection)
+		iter, err = projection.ProjectionIterator(iter, closer, params.Projection)
 		if err != nil {
 			return lazyerrors.Error(err)
 		}
